+++ tests/testthat/test-writeManifest.R
@@ -67,10 +67,10 @@ test_that("Rmd without a python block doesn't include reticulate or python in th
 # Quarto Tests
 
 test_that("Quarto website includes quarto in the manifest", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("quarto-website-r")
-  manifest <- makeManifest(appDir, quarto = quarto)
+  manifest <- makeManifest(appDir, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "knitr")
@@ -78,11 +78,11 @@ test_that("Quarto website includes quarto in the manifest", {
 })
 
 test_that("Quarto document includes quarto in the manifest", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("quarto-doc-none")
   appPrimaryDoc <- "quarto-doc-none.qmd"
-  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = quarto)
+  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "markdown")
@@ -90,11 +90,11 @@ test_that("Quarto document includes quarto in the manifest", {
 })
 
 test_that("Specifying quarto arg includes quarto in the manifest, even with no appPrimaryDoc specified (.qmd)", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("quarto-doc-none")
   appPrimaryDoc <- NULL
-  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = quarto)
+  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "markdown")
@@ -102,11 +102,11 @@ test_that("Specifying quarto arg includes quarto in the manifest, even with no a
 })
 
 test_that("Specifying quarto arg includes quarto in the manifest, even with no appPrimaryDoc specified (.Rmd)", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("shiny-rmds")
   appPrimaryDoc <- NULL
-  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = quarto)
+  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-shiny")
   expect_equal(manifest$quarto$engines, "knitr")
@@ -114,20 +114,20 @@ test_that("Specifying quarto arg includes quarto in the manifest, even with no a
 })
 
 test_that("specifying quarto arg with non-quarto app does not include quarto in the manifest", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("shinyapp-singleR")
   appPrimaryDoc <- "single.R"
-  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = quarto)
+  manifest <- makeManifest(appDir, appPrimaryDoc, quarto = TRUE)
 
   expect_null(manifest$quarto)
 })
 
 test_that("Quarto shiny project includes quarto in the manifest", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
   appDir <- test_path("quarto-proj-r-shiny")
-  manifest <- makeManifest(appDir, quarto = quarto)
+  manifest <- makeManifest(appDir, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-shiny")
   expect_equal(manifest$quarto$engines, "knitr")
@@ -136,11 +136,11 @@ test_that("Quarto shiny project includes quarto in the manifest", {
 
 test_that("Quarto R + Python website includes quarto and python in the manifest", {
   skip_if_not_installed("reticulate")
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
   python <- pythonPathOrSkip()
 
   appDir <- test_path("quarto-website-r-py")
-  manifest <- makeManifest(appDir, python = python, quarto = quarto)
+  manifest <- makeManifest(appDir, python = python, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "knitr")
@@ -152,11 +152,12 @@ test_that("Quarto R + Python website includes quarto and python in the manifest"
 
 test_that("Quarto Python-only website gets correct manifest data", {
   skip_if_not_installed("reticulate")
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
+
   python <- pythonPathOrSkip()
 
   appDir <- test_path("quarto-website-py")
-  manifest <- makeManifest(appDir, python = python, quarto = quarto)
+  manifest <- makeManifest(appDir, python = python, quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "jupyter")
@@ -167,24 +168,17 @@ test_that("Quarto Python-only website gets correct manifest data", {
   expect_null(manifest$packages)
 })
 
-test_that("Deploying a Quarto project without Quarto info in an error", {
+test_that("Deploying a Quarto project without Quarto is an error", {
+  local_mocked_bindings(quarto_path = function() NULL)
+
   appDir <- test_path("quarto-website-r")
   expect_snapshot(makeManifest(appDir, quarto = NULL), error = TRUE)
 })
 
-test_that("Deploying a Quarto doc without Quarto info in an error", {
-  appDir <- test_path("quarto-doc-none")
-  appPrimaryDoc <- "quarto-doc-none.qmd"
-  expect_snapshot(
-    makeManifest(appDir, appPrimaryDoc = appPrimaryDoc),
-    error = TRUE
-  )
-})
-
 test_that("Deploying R Markdown content with Quarto gives a Quarto app mode", {
-  quarto <- quartoPathOrSkip()
+  skip_if_no_quarto()
 
-  manifest <- makeManifest(test_path("test-rmds"), "simple.Rmd", quarto = quarto)
+  manifest <- makeManifest(test_path("test-rmds"), "simple.Rmd", quarto = TRUE)
 
   expect_equal(manifest$metadata$appmode, "quarto-static")
   expect_equal(manifest$quarto$engines, "knitr")